Break-out functions into files, make a class
